free_tier_model_test/requirements.txt
flask
python-dotenv
google-genai
pillow
requests
